+++ src/lib/messages/bindings/Countersigner.ts
@@ -1,8 +1,8 @@
 import { ObjectIdentifier, OctetString } from 'asn1js';
-import { serializeSequence } from '../../../asn1';
-import { SignedData } from '../../../crypto_wrappers/cms/signedData';
-import Certificate from '../../../crypto_wrappers/x509/Certificate';
-import { RELAYNET_OIDS } from '../../../oids';
+import { serializeSequence } from '../../asn1';
+import { SignedData } from '../../crypto_wrappers/cms/signedData';
+import Certificate from '../../crypto_wrappers/x509/Certificate';
+import { RELAYNET_OIDS } from '../../oids';
 
 /**
  * Utility to sign and verify countersignatures.
